rename and move mp ru helper file
